# collection-postman/swagger.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "example-junit-test",
    "description": "API desenvolvida para servir de exemplo de implementação de securança com spring security com token JWT e documentação com SpringDoc.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://springdoc.org/"
    },
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Clientes",
      "description": "Responsável por gerenciar o cadastro de clientes."
    },
    {
      "name": "Login",
      "description": "Responsável por autenticar o usuário na API."
    }
  ],
  "paths": {
    "/v1/produtos": {
      "post": {
        "tags": [
          "produto-controller"
        ],
        "operationId": "save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ordens-compra": {
      "post": {
        "tags": [
          "ordem-compra-controller"
        ],
        "operationId": "save_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdemCompraInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrdemCompraDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "Login do usuário",
        "description": "Responsável por fazer login e autenticar o usuário.",
        "operationId": "login",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "senha",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/estoques": {
      "post": {
        "tags": [
          "estoque-controller"
        ],
        "operationId": "save_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstoqueInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EstoqueDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/clientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Lista os clientes",
        "description": "Responsável por listar todos os clientes cadastrados",
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clientes"
        ],
        "summary": "Salva um cliente",
        "description": "Responsável por salvar um cliente",
        "operationId": "save_3",
        "requestBody": {
          "description": "Representação de um novo cliente",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/produtos/{id}": {
      "get": {
        "tags": [
          "produto-controller"
        ],
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/clientes/{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "summary": "Busca um cliente pelo ID",
        "description": "Responsável por retornar um cliente cadastrado pelo ID",
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do cliente",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "404"
          },
          "message": {
            "type": "string",
            "example": "Recurso com id 1, não encontrado."
          },
          "detail": {
            "type": "string"
          }
        }
      },
      "ProdutoInput": {
        "type": "object",
        "properties": {
          "descricao": {
            "type": "string"
          },
          "preco": {
            "type": "number"
          },
          "porcentagemDesconto": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ProdutoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "descricao": {
            "type": "string"
          },
          "preco": {
            "type": "number"
          },
          "porcentagemDesconto": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CarrinhoComprasInput": {
        "type": "object",
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemCarrinhoComprasInput"
            }
          }
        }
      },
      "ItemCarrinhoComprasInput": {
        "type": "object",
        "properties": {
          "produtoId": {
            "type": "integer",
            "format": "int64"
          },
          "quantidade": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "OrdemCompraInput": {
        "type": "object",
        "properties": {
          "clienteId": {
            "type": "integer",
            "format": "int64"
          },
          "formaPagamento": {
            "type": "string",
            "enum": [
              "BOLETO",
              "PIX",
              "CARTAO_CREDITO",
              "CARTAO_DEBITO"
            ]
          },
          "carrinhoCompras": {
            "$ref": "#/components/schemas/CarrinhoComprasInput"
          },
          "enderecoEntregaId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CarrinhoComprasDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemCarrinhoComprasDto"
            }
          },
          "valorTotal": {
            "type": "number"
          }
        }
      },
      "EnderecoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "cep": {
            "type": "string"
          },
          "logradouro": {
            "type": "string"
          },
          "complemento": {
            "type": "string"
          },
          "localidade": {
            "type": "string"
          },
          "bairro": {
            "type": "string"
          },
          "uf": {
            "type": "string"
          }
        }
      },
      "ItemCarrinhoComprasDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "quantidade": {
            "type": "integer",
            "format": "int32"
          },
          "produto": {
            "type": "string"
          }
        }
      },
      "OrdemCompraDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "carrinhoCompras": {
            "$ref": "#/components/schemas/CarrinhoComprasDto"
          },
          "formaPagamento": {
            "type": "string",
            "enum": [
              "BOLETO",
              "PIX",
              "CARTAO_CREDITO",
              "CARTAO_DEBITO"
            ]
          },
          "enderecoEntrega": {
            "$ref": "#/components/schemas/EnderecoDto"
          },
          "cliente": {
            "type": "string"
          }
        }
      },
      "TokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "EstoqueInput": {
        "type": "object",
        "properties": {
          "produtoId": {
            "type": "integer",
            "format": "int64"
          },
          "quantidade": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "EstoqueDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "quantidade": {
            "type": "integer",
            "format": "int32"
          },
          "produto": {
            "type": "string"
          }
        }
      },
      "CartaoInput": {
        "type": "object",
        "properties": {
          "bandeira": {
            "type": "string",
            "enum": [
              "MASTERCARD",
              "VISA",
              "ELO",
              "AMERICAN_EXPRESS",
              "HIPERCARD"
            ]
          },
          "numero": {
            "type": "string"
          },
          "validade": {
            "type": "string"
          },
          "cvv": {
            "type": "string"
          }
        }
      },
      "ClienteInput": {
        "required": [
          "cpf",
          "email",
          "nome",
          "telefone"
        ],
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "enderecos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnderecoInput"
            }
          },
          "cartoes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartaoInput"
            }
          }
        }
      },
      "EnderecoInput": {
        "type": "object",
        "properties": {
          "cep": {
            "type": "string"
          },
          "logradouro": {
            "type": "string"
          },
          "complemento": {
            "type": "string"
          },
          "localidade": {
            "type": "string"
          },
          "bairro": {
            "type": "string"
          },
          "uf": {
            "type": "string"
          }
        }
      },
      "ClienteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "enderecos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnderecoDto"
            }
          }
        }
      }
    }
  }
}